requirements.txt
hikari
hikari-lightbulb
uvloop>= 0.16; sys_platform != "win32"
python_dateutil
python-decouple
twitchAPI
aiofiles
pyngrok
tomli >= 1.1.0 ; python_version < "3.11"
